Updated eslint

+++ JS/.eslintrc.js
@@ -194,7 +194,7 @@ module.exports = {
     'prefer-template': 'error',
     'quotes': [
       'error',
-      'double',
+      'single',
       {
         avoidEscape: true,
         allowTemplateLiterals: true,
